Extend IDataRepo to work with Users
