+++ warhammer-core/WarhammerCore.Abstract/Interfaces/IDataRepo.cs
@@ -1,6 +1,8 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using WarhammerCore.Abstract.Models;
+using WarhammerCore.Abstract.Models.User;
 
 namespace WarhammerCore.Abstract.Interfaces
 {
@@ -9,8 +11,10 @@ namespace WarhammerCore.Abstract.Interfaces
     /// </summary>
     public interface IDataRepo
     {
+        #region Profession
+
         /// <summary>
-        /// Get list of all properties IDs. 
+        /// Get list of all properties IDs.
         /// </summary>
         Task<IEnumerable<string>> GetProfessionsAsync();
 
@@ -19,5 +23,31 @@ namespace WarhammerCore.Abstract.Interfaces
         /// </summary>
         /// <param name="professionId">Unique id of the profession.</param>
         Task<Profession> GetProfessionAsync(string professionId);
+
+        #endregion Profession
+
+        #region User
+
+        /// <summary>
+        /// Create user in the database.
+        /// </summary>
+        /// <param name="userInfo">User details from registration page.</param>
+        /// <returns>ID of the user.</returns>
+        Task<string> CreateUserAsync(string email, string password);
+
+        /// <summary>
+        /// Get the user by his/her unique ID.
+        /// </summary>
+        /// <param name="userId">Unique user ID (GUID).</param>
+        /// <returns>User account information or an EMPTY user when user was not found in the database.</returns>
+        Task<UserInfo> GetUserByIdAsync(string userId);
+
+        /// <summary>
+        /// Get the user by his/her unique mail address.
+        /// </summary>
+        /// <returns>User account information or an EMPTY user when user was not found in the database.</returns>
+        Task<UserInfo> GetUserByEmailAsync(string email);
+
+        #endregion User
     }
 }
